Make stdlib files not use the file extension when requiring
This has the potential to break some tests for stdlib as it will actually be loading
the compiled files now.

This commit only includes the changes which should be safe to change; that is, changes
which do not alter the logic of the code. There are some other places that need to be
changed but the code performs checks for file existence or depend on the file extension
being present for some other reason. The more sensitive libraries may have to be gone
through one piece at a time.