Remove xcode_support from the standard rakefile installation
This change introduces an install:standard task which is now
the default. It will still install the necessary components
so that the Xcode integration script can be run, but it will
not run the integration install script.

Xcode integration will still be installed by default when
using the package installer, and the script can be automatically
run using the install:all task.