Remove obsoleted lib/objc_ext extensions
There are no project using these extensions on
Github (at the time of this commit), so it is a
safe gamble that nobody will miss them.

Half of the functionality given here is already
provided by BridgeSupport and the other half is
trivial for a user to add themselves if they
really want it.